Split up structures into separate files, keeping backwards compatibility with previous import style though
