+++ src/structures/__init__.py
@@ -1,2 +1,5 @@
-from .structures import *
+from .colour import Colour, WHITE, BLACK, RED, GREEN, BLUE, TRANSPARENT
+from .dynamic_colour import DynamicColour
+from .point import Point
+from .size import Size
 from .factory import gen_heap, gen_list, gen_set
