+++ src/utility/compatibility.py
@@ -1,6 +1,6 @@
 try:
     from typing import Final
-except ImportError: # work-around for Python < 3.8
+except ImportError:  # work-around for Python < 3.8
     from typing import Tuple
     Final = Tuple
